backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.6.0
future==1.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-social-auth==0.3.6
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
serializers==0.2.4
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
